Enable R/O ops concurrency with other R/O ops.
This change adds some new guarantees. Selected R/O operations are now
guaranteed to be safe for concurrent access with other selected R/O
operations, where R/O is with respect to mutation of the DB.  Formerly
some read only operations mutated internal in-memory objects (eg. a disk
cache, the LRU list, ...) without a mutex. That was okay because those
operation were never declared to be safe for concurrent access by
multiple goroutines.

The particular motivation for this change is the cznic/ql#100 issue
which reports a race condition  when, erroneously, allocator Get and
B+Tree Seek were called concurrently without proper synchronization.

LLDB clients need no more to care about the synchronization of the
selected R/O operations as long as the client does not concurrently
mutate the DB.